backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.16,<2.0
pymysql>=1.1,<2.0
pydantic-settings>=2.10,<3.0
python-jose>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt<4.1
httpx>=0.28,<1.0
pytest>=8.3,<9.0
cryptography>=46.0,<47.0
python-multipart>=0.0.9,<1.0
Pillow>=11.2,<12.0
